Refactoring Simulator v1.2; Adding BigInteger Conversion

+++ src/main/java/com/sparta/engineering72/Animal/Fox/MaleFox.java
@@ -2,20 +2,22 @@ package com.sparta.engineering72.Animal.Fox;
 
 import com.sparta.engineering72.Animal.Animal;
 
+import java.math.BigInteger;
+
 public class MaleFox extends Fox{
-    private long count;
+    private BigInteger count;
 
     public MaleFox() {
         super(Animal.Gender.MALE);
         this.age = 0;
-        this.count = 1;
+        this.count = BigInteger.valueOf(1);
     }
 
-    public long getCount() {
+    public BigInteger getCount() {
         return count;
     }
 
-    public void setCount(long count) {
+    public void setCount(BigInteger count) {
         this.count = count;
     }
 }
